[bugfix] fix NativeBroker.defragXMLResource
A document that is frequently modified using update expressions will at
some point trigger defragmentation.
This PR fixes a NullPointer and a PermissionDeniedException that are
thrown when NativeBroker.defragXMLResource is called.
The modified document would be removed from all indexes it was part of
and also not be queriable anymore.